+++ NSArrayAdditions.m
@@ -48,7 +48,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
 - (NSArray *)AE_arrayWithUniqueObjects
 {
 	NSMutableArray *const array = [[self mutableCopy] autorelease];
-	unsigned i = 0, count;
+	NSUInteger i = 0, count;
 	for(; i < (count = [array count]); i++) [array removeObject:[array objectAtIndex:i] inRange:NSMakeRange(i + 1, count - i - 1)];
 	return array;
 }
